Major update - Changed: Moved all guild shop references into a new file called guild_shop.py
- Added:
Guild Shop Products Trials and Active Trials in both types/guild.py and guild_shop.py
New get_guild_subscriptions_trials method in HTTPClient
